IrisSegmentIO/requirements.txt
Django>=4.0,<5.0
tensorflow>=2.0,<3.0
opencv-python-headless>=4.5.0,<5.0
numpy>=1.21.0,<2.0
Pillow>=8.0,<10.0